requirements.txt
pillow==11.0.0
pycryptodome==3.21.0
